changed webapp.py to no longer use `group_politician_similarity` table
This fixes two bugs:
- a bug Didier Deshommes reported, in which some politicians display
  with agreement information for only one interest group, even though
  many other interest groups may have expressed interest on a smaller
  number of bills;
- a bug in which interest groups who disagreed with a politician on
  every issue were not displayed at all.

It leaves a bug in which politicians who abstained from a vote are
counted as disagreeing with all interest groups on the issue.

Additionally, this speeds up page load times by a few hundred
milliseconds, because the `position` table is properly indexed to
support efficient query execution, while the
`group_politician_similarity` table was not; it permits the removal
of all of the code to create and populate `group_politician_similarity`.